ERD Schema
The schema is based more on joins rather than on PK and FK relationships. 
Eight rows from SalesBySKU table were deleted to create a FK relationship with the Products table's primary key (SKU). This looked like an unnecessary action later on. However as the  SalesBySKU was not used for the queries later, it was not recreated.